+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added `ed2k` hash support for eDonkey/eMule/MLDonkey-compatible file hashes.
+- Added `tiger` Tiger Tree Hash support for Gnutella2/Direct Connect-compatible file hashes.
 
 ## 0.4.0 - 2026-04-30